+++ pmd/src/site/markdown/changelog.md
@@ -58,6 +58,7 @@
 * Fixed [bug 1141]: ECMAScript: getFinallyBlock() is buggy.
 * Fixed [bug 1142]: ECMAScript: getCatchClause() is buggy.
 * Fixed [bug 1144]: CPD encoding argument has no effect
+* Fixed [bug 1045]: //NOPMD not working (or not implemented) with ECMAscript
 * Fixed [bug 1146]: UseArrayListInsteadOfVector false positive when using own Vector class
 * Fixed [bug 1147]: EmptyMethodInAbstractClassShouldBeAbstract false positives
 * Fixed [bug 1150]: "EmptyExpression" for valid statements!
@@ -85,6 +86,7 @@
 [bug 1141]: https://sourceforge.net/p/pmd/bugs/1141
 [bug 1142]: https://sourceforge.net/p/pmd/bugs/1142
 [bug 1144]: https://sourceforge.net/p/pmd/bugs/1144
+[bug 1045]: https://sourceforge.net/p/pmd/bugs/1045
 [bug 1146]: https://sourceforge.net/p/pmd/bugs/1146
 [bug 1147]: https://sourceforge.net/p/pmd/bugs/1147
 [bug 1150]: https://sourceforge.net/p/pmd/bugs/1150